S104 split: extract provisioner_db.py — main.py now 482 lines (under 500)
Final S104 extraction. main.py shrunk from 3,439 to 482 lines; all 718
provisioner tests pass; SonarCloud S104 should now report 0 findings on
shifter/engine/provisioner/main.py.

Sibling modules created during the split:
- state_helpers.py (218 lines): pure validation + state builders
- ngfw_runtime_ops.py (175 lines): NGFW start/stop runtime ops
- instance_setup.py (884 lines): per-instance + DC + polaris setup
- terraform_ops.py (649 lines): terraform provision/destroy + variables
- ngfw_runtime.py (574 lines): NGFW polling/config/routes/state writes
- provisioner_db.py (500 lines): DB access + range/NGFW state lookups
- main.py (482 lines): get_ami_id, get_agent_presigned_url, DynamicPlan,
  the __main__ argparse entrypoint, and explicit re-exports so existing
  patch("main.X") test mocks continue to work without test changes.
